久しぶり

+++ math/prime-table.hpp
@@ -1,9 +1,9 @@
 #include "math/prime.hpp"
 
-vector< int >  prime_table(int n) {
+vector< ll >  prime_table(int n) {
     if(n <= 1) return {};
     auto d = prime(n);
-    vector< int > primes;
+    vector< ll > primes;
     rep(i, d.size()) {
         if(d[i]) primes.push_back(i);
     }
